pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evolution-engine"
version = "0.1.0"
description = "Evolution Engine V1 (Observer->Actor->Judge->Router)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.49.0",
  "openai>=1.30.0",
  "PyGithub>=2.0.0",
  "unidiff>=0.7.0",
  "click>=8.1.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
evolve = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "cli",
  "config_loader",
  "github_client",
  "llm",
  "observer",
  "actor",
  "judge",
  "router",
  "hard_stops",
  "history",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
